DEVOPS-12554 useadderror property windows

+++ src/RabbitMQClient.cpp
@@ -349,7 +349,7 @@ void RabbitMQClient::basicConsumeMessageImpl(Biterp::CallContext& ctx) {
 		messageQueue.pop();
 	}
 	ctx.setStringResult(u16Converter.from_bytes(lastMessage.body), outdata);
-	ctx.setIntResult(lastMessage.messageTag, outMessageTag);
+	ctx.setIntResult((int)lastMessage.messageTag, outMessageTag);
 	ctx.setBoolResult(true);
 }
